fix: Make streaming-schedule and inference more robust
This new, default inference program aims to record fewer checkpoints
and ensure the streaming schedule doesn't choke on partial columns
that don't have any values observed.

It is also simpler than what we had before and I (@schaechtle) hopes
that it is thus more suitable as a starting point for users.

Fixes #17