refactor: Factor out the recording of the maximum number of views of all models in the ensemble
This number is needed to merge the individual SPPL models into an ensemble (`merged.json`). This is necessary because all children of the top node need to generate the same variables.

Before, this was done as a side effect of generating the dep-prob.jon and, confusingly, saved as a field in dep-prob.json; alongside the columns.