docs: Restructured quick start guide
The structure-learning document was redundant with the quick start guide. I pulled out the common elements into individual files, and referenced those files from within the structure learning and quick start docs. The population-assembly doc was renamed to data-preparation. Some additional testing needs to be done once this is uploaded - I was unable to test the embedded without the docs being hosted. A proper intro for structure learning in the structure-learning doc also still needs to be written. There is a placeholder there now.